+++ jetson/percep_obs_detect/CMakeLists.txt
@@ -16,8 +16,8 @@ find_package(OpenGL REQUIRED)
 find_package(glfw3 REQUIRED)
 message("${GLFW_INCLUDE_DIR}")
 include_directories(SYSTEM ${GLFW_INCLUDE_DIR} ${GLEW_INCLUDE_DIRS})
-include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
-file(GLOB SOURCE_FILES CONFIGURE_DEPENDS "src/*.cpp")
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_SOURCE_DIR}/include/imgui)
+file(GLOB_RECURSE SOURCE_FILES CONFIGURE_DEPENDS "src/*.cpp")
 add_executable (viewer ${SOURCE_FILES})
 set_target_properties(viewer PROPERTIES LINKER_LANGUAGE CXX)
